Moved glue.qt.widgets.style_dialog to glue.core.qt.style_dialog
