+++ glue/core/qt/tests/test_style_dialog.py
